Cleaning out old code for push to master
